utils: reorganized c++ code to standard directory structure
