+++ .gitignore
@@ -1,2 +1,3 @@
 .vscode/tasks.json
 .vscode/launch.json
+cpp/src/test/*_test
