+++ cpp/include/bounded_queue.h
