+++ cpp/include/synchronized_queue.h
