+++ cpp/tests/queue_test_perf.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <chrono>
 #include <queue>
-#include "bounded_queue.h"
+#include "../include/bounded_queue.h"
 
 using namespace efficient;
 
@@ -14,13 +14,13 @@ int main(int argc, char** argv) {
     }
     auto elapsed = std::chrono::high_resolution_clock::now() - start;
     long long microseconds = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
-    cout << "Bounded queue took " << microseconds << " microseconds to push " << QUEUE_MAX_SIZE_DEFAULT << " elements" << endl;
+    std::cout << "Bounded queue took " << microseconds << " microseconds to push " << QUEUE_MAX_SIZE_DEFAULT << " elements" << std::endl;
 
     start = std::chrono::high_resolution_clock::now();
     q.clear();
     elapsed = std::chrono::high_resolution_clock::now() - start;
     microseconds = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
-    cout << "Bounded queue took " << microseconds << " microseconds to clear" << endl;
+    std::cout << "Bounded queue took " << microseconds << " microseconds to clear" << std::endl;
 
     std::queue<int> sq;
     start = std::chrono::high_resolution_clock::now();
@@ -29,7 +29,7 @@ int main(int argc, char** argv) {
     }
     elapsed = std::chrono::high_resolution_clock::now() - start;
     microseconds = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
-    cout << "STL queue took " << microseconds << " microseconds to push " << QUEUE_MAX_SIZE_DEFAULT << " elements" << endl;
+    std::cout << "STL queue took " << microseconds << " microseconds to push " << QUEUE_MAX_SIZE_DEFAULT << " elements" << std::endl;
 
     start = std::chrono::high_resolution_clock::now();
     while (!sq.empty()) {
@@ -37,5 +37,5 @@ int main(int argc, char** argv) {
     }
     elapsed = std::chrono::high_resolution_clock::now() - start;
     microseconds = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
-    cout << "STL queue took " << microseconds << " microseconds to clear" << endl;
+    std::cout << "STL queue took " << microseconds << " microseconds to clear" << std::endl;
 }
